refactor: style 전용 인터페이스 분리 및 공통 컴포넌트 import 경로 수정정
